REFACTOR: reorganize files to database directory
