+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve('api/config', 'database.js'),
-  'models-path': path.resolve('api', 'models'),
-  'seeders-path': path.resolve('api', 'seeders'),
-  'migrations-path': path.resolve('api', 'migrations')
+  'config': path.resolve('api/database/config', 'database.js'),
+  'models-path': path.resolve('api/database', 'models'),
+  'seeders-path': path.resolve('api/database', 'seeders'),
+  'migrations-path': path.resolve('api/database', 'migrations')
 };
