+++ api/__test__/unit/models/user.test.js
@@ -1,5 +1,5 @@
 process.env.NODE_ENV = 'test'
-const db = require('../../../models')
+const db = require('../../../database/models')
 
 describe('Testing User model', () => {
