+++ api/services/PetService.js
@@ -1,6 +1,6 @@
 const Services = require('./Services')
 const UserService = require('./UserService.js')
-const dataSource = require('../models')
+const dataSource = require('../database/models')
 const { v4: uuid } = require('uuid')
 
 const userService = new UserService()
